+++ src/run.jl
@@ -1,9 +1,33 @@
+@doc """
+Run code chunks and capture output from parsed document.
 
-
-
+```julia
 function run(doc::WeaveDoc; doctype = "pandoc", plotlib="Gadfly", informat="noweb",
-        out_path=:doc, fig_path = "figures", fig_ext = nothing, cache_path = "cache")
+    out_path=:doc, fig_path = "figures", fig_ext = nothing,
+    cache_path = "cache", cache = :off)
+```
+
+* `doctype`: see `list_out_formats()`
+* `plotlib`: `"PyPlot"`, `"Gadfly"`, or `"Winston"`
+* `informat`: `"noweb"` of `"markdown"`
+* `out_path`: Path where the output is generated. Can be: `:doc`: Path of the source document, `:pwd`: Julia working directory,
+`"somepath"`: Path as a string e.g `"/home/mpastell/weaveout"`
+* `fig_path`: where figures will be generated, relative to out_path
+* `fig_ext`: Extension for saved figures e.g. `".pdf"`, `".png"`. Default setting depends on `doctype`.
+* `cache_path`: where of cached output will be saved.
+* `cache`: controls caching of code: `:off` = no caching, `:all` = cache everything,
+ `:user` = cache based on chunk options, `:refresh`, run all code chunks and save new cache.
+
+**Note:** Run command from terminal and not using IJulia, Juno or ESS, they tend to mess with capturing output.
+""" ->
+function Base.run(doc::WeaveDoc; doctype = "pandoc", plotlib="Gadfly", informat="noweb",
+        out_path=:doc, fig_path = "figures", fig_ext = nothing,
+        cache_path = "cache", cache = :off)
+    #cache :all, :user, :off, :refresh
+
+
     doc.cwd = get_cwd(doc, out_path)
+    doc.doctype = doctype
     doc.format = formats[doctype]
     set_rc_params(doc.format.formatdict, fig_path, fig_ext)
 
@@ -18,17 +42,57 @@ function run(doc::WeaveDoc; doctype = "pandoc", plotlib="Gadfly", informat="nowe
     report = Report(doc.cwd, doc.basename, doc.format.formatdict)
     pushdisplay(report)
 
+    if cache != :off || cache != :refresh
+        cached = read_cache(doc, cache_path)
+        cached == nothing && info("No cached results found, running code")
+    else
+        cached = nothing
+    end
+
     executed = Any[]
-    for chunk in copy(doc.chunks)
-        result_chunk = eval_chunk(chunk, report, SandBox)
+    n = length(doc.chunks)
+
+    for i = 1:n
+        chunk = doc.chunks[i]
+        result_chunk = run_chunk(chunk, report, SandBox, cached, cache, i)
         push!(executed, result_chunk)
     end
+
     popdisplay(report)
+
+    if cache != :off
+        write_cache(doc, cache_path)
+    end
 
     #Clear variables from used sandbox
     clear_sandbox(SandBox)
     doc.chunks = executed
     return doc
+end
+
+function run_chunk(chunk::CodeChunk, report::Report, SandBox::Module, cached, cache, idx)
+    defaults = copy(rcParams[:chunk_defaults])
+    options = copy(chunk.options)
+    try
+        options = merge(rcParams[:chunk_defaults], options)
+    catch
+        options = rcParams[:chunk_defaults]
+        warn("Invalid format for chunk options line: $(chunk.start_line)")
+    end
+
+    merge!(chunk.options, options)
+
+    delete!(chunk.options, :options)
+
+    if cached != nothing && (cache == :all || (cache ==:user && chunk.options[:cache]))
+        result_chunk = restore_chunk(chunk, cached, idx)
+    else
+        result_chunk = eval_chunk(chunk, report, SandBox)
+    end
+end
+
+function run_chunk(chunk::DocChunk, report::Report, SandBox::Module, cached, cache, idx)
+    return chunk
 end
 
 function run_block(code_str, report::Report, SandBox::Module)
@@ -89,18 +153,6 @@ end
 
 function eval_chunk(chunk::CodeChunk, report::Report, SandBox::Module)
     info("Weaving chunk $(chunk.number) from line $(chunk.start_line)")
-    defaults = copy(rcParams[:chunk_defaults])
-    options = copy(chunk.options)
-    try
-        options = merge(rcParams[:chunk_defaults], options)
-    catch
-        options = rcParams[:chunk_defaults]
-        warn("Invalid format for chunk options line: $(chunk.start_line)")
-    end
-
-    merge!(chunk.options, options)
-    #delete!(chunk.options, :options)
-    #@show chunk.options
 
     if !chunk.options[:eval]
         chunk.output = ""
@@ -108,11 +160,13 @@ function eval_chunk(chunk::CodeChunk, report::Report, SandBox::Module)
         return chunk
     end
 
+
     report.fignum = 1
     report.cur_result = ""
     report.figures = String[]
     report.cur_chunk = chunk
     report.term_state = :text
+
     if haskey(report.formatdict, :out_width) && chunk.options[:out_width] == nothing
         chunk.options[:out_width] = report.formatdict[:out_width]
     end
@@ -136,6 +190,9 @@ function eval_chunk(chunk::DocChunk, report::Report, SandBox)
     chunk
 end
 
+
+
+
 #Set all variables to nothing
 function clear_sandbox(SandBox::Module)
     for name = names(SandBox, true)
@@ -148,7 +205,7 @@ end
 
 function get_figname(report::Report, chunk; fignum = nothing)
     figpath = joinpath(report.cwd, chunk.options[:fig_path])
-    isdir(figpath) || mkdir(figpath)
+    isdir(figpath) || mkpath(figpath)
     ext = chunk.options[:fig_ext]
     fignum == nothing && (fignum = report.fignum)